Compatibility fixes

+++ src/stdlib/numX.h
@@ -72,7 +72,7 @@ MACROLIKE CONSTFUNC NUM_T NAMESPACED(from_int)(Int_t i, bool truncate) {
         NUM_T ret = mpz_get_d(*i.big);
         if (!truncate) {
             mpz_t roundtrip;
-            mpz_init_set_d(roundtrip, ret);
+            mpz_init_set_d(roundtrip, (double)ret);
             if unlikely (mpz_cmp(*i.big, roundtrip) != 0)
                 fail("Could not convert integer to " TYPE_STR " without losing precision: ", i);
         }
